# db/eliasdb/cli/swagger.yaml
basePath: /db
definitions:
  Error:
    description: A human readable error mesage.
    type: string
...
host: 127.0.0.1:9090
info:
  description: Query and modify the EliasDB datastore.
  title: EliasDB API
  version: 1.0.0
paths:
...
  /ecal:
    delete:
      description: The ecal endpoint forwards web requests to the ECAL backend.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A result object generated by ECAL scripts.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Forward web requests to the ECAL backend.
    get:
      description: The ecal endpoint forwards web requests to the ECAL backend.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A result object generated by ECAL scripts.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Forward web requests to the ECAL backend.
    post:
      description: The ecal endpoint forwards web requests to the ECAL backend.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A result object generated by ECAL scripts.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Forward web requests to the ECAL backend.
    put:
      description: The ecal endpoint forwards web requests to the ECAL backend.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A result object generated by ECAL scripts.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Forward web requests to the ECAL backend.
...
  /v1/cluster:
    get:
      description: >-
        The cluster endpoint returns the cluster state info which contains
        cluster members and their state.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A key-value map.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Return cluster specific information.
  /v1/cluster/log:
    delete:
      description: >-
        A delete call to the log endpoint resets the cluster related log and
        clears the ring buffer in memory.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: Cluster related log was reset.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Reset the cluster log.
    get:
      description: >-
        The cluster log endpoint returns the latest cluster related log messages
        from a memory ring buffer.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A list of log messages.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Return latest cluster related log messages.
  /v1/cluster/memberinfos:
    get:
      description: >-
        The memberinfos returns the static member info of every known cluster
        member. If a member is not reachable its info contains a single
        key-value pair with the key error and an error message as value.
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: A map of memberinfos (keys are member names).
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Return static member info of every known cluster member.
  /v1/cluster/{command}:
    put:
      consumes:
        - application/json
      description: The cluster can be controlled via this command endpoint on any member.
      parameters:
        - description: 'Valid commands are: ping, join and eject.'
          in: path
          name: command
          required: true
          type: string
        - description: Arguments for a command
          in: body
          name: args
          required: true
          schema:
            properties:
              name:
                description: >-
                  Name of a cluster member (ping/join=member to contact,
                  eject=member to eject).
                type: string
              netaddr:
                description: >-
                  Network address of a member e.g. localhost:9030
                  (ping/join=member address to contact)
                type: string
            type: object
      produces:
        - text/plain
        - application/json
      responses:
        '200':
          description: >-
            Only the ping command returns its result. All other positive
            responses are empty.
        default:
          description: Error response
          schema:
            $ref: '#/definitions/Error'
      summary: Commands can be given to the cluster by using PUT requests.
...
produces:
  - application/json
schemes:
  - https
swagger: '2.0'